TG-17: Added validations from id in path params- Added Not found error response.  Fixed successful update response.

+++ src/controllers/projects.js
@@ -1,13 +1,30 @@
 import Project from '../models/Projects';
 
+const { ObjectId } = require('mongoose').Types;
+
+const isValidId = (id) => {
+  try {
+    const oid = new ObjectId(id);
+    return ObjectId.isValid(id)
+    && oid.toString() === id;
+  } catch {
+    return false;
+  }
+};
 const getProjectById = async (req, res) => {
-  if (req.params.id) {
+  if (req.params.id && isValidId(req.params.id)) {
     try {
       const retrievedProject = await Project.findById(req.params.id);
-      res.status(200).json({
-        message: `Project with id=${req.params.id} found.`,
-        data: retrievedProject,
-      });
+      if (retrievedProject !== null) {
+        res.status(200).json({
+          message: `Project with id=${req.params.id} found.`,
+          data: retrievedProject,
+        });
+      } else {
+        res.status(404).json({
+          message: `Project with id=${req.params.id} not found.`,
+        });
+      }
     } catch (error) {
       res.status(500).json({
         message: error.message,
@@ -15,19 +32,26 @@ const getProjectById = async (req, res) => {
     }
   } else {
     res.status(400).json({
-      message: 'Error: no project id provided',
+      message: 'Error: id is invalid or missing.',
     });
   }
 };
 
 const updateProjectById = async (req, res) => {
-  if (req.params.id) {
+  if (req.params.id && isValidId(req.params.id)) {
     try {
-      const updatedProject = await Project.findByIdAndUpdate(req.params.id, req.body);
-      res.status(200).json({
-        message: `Project with id=${req.params.id} deleted.`,
-        data: updatedProject,
-      });
+      const updateResponse = await Project.findByIdAndUpdate(req.params.id, req.body);
+      if (updateResponse !== null) {
+        const updatedProject = await Project.findById(req.params.id);
+        res.status(200).json({
+          message: `Project with id=${req.params.id} has been updated.`,
+          data: updatedProject,
+        });
+      } else {
+        res.status(404).json({
+          message: `Project with id=${req.params.id} not found.`,
+        });
+      }
     } catch (error) {
       res.status(500).json({
         message: error.message,
@@ -35,18 +59,24 @@ const updateProjectById = async (req, res) => {
     }
   } else {
     res.status(400).json({
-      message: 'Error: no project id provided',
+      message: 'Error: id is invalid or missing.',
     });
   }
 };
 
 const deleteProjectById = async (req, res) => {
-  if (req.params.id) {
+  if (req.params.id && isValidId(req.params.id)) {
     try {
-      await Project.findByIdAndDelete(req.params.id);
-      res.status(202).json({
-        message: `Project with id=${req.params.id} deleted.`,
-      });
+      const response = await Project.findByIdAndDelete(req.params.id);
+      if (response !== null) {
+        res.status(202).json({
+          message: `Project with id=${req.params.id} deleted.`,
+        });
+      } else {
+        res.status(404).json({
+          message: `Project with id=${req.params.id} not found.`,
+        });
+      }
     } catch (error) {
       res.status(500).json({
         message: error.message,
@@ -54,7 +84,7 @@ const deleteProjectById = async (req, res) => {
     }
   } else {
     res.status(400).json({
-      message: 'Error: no project id provided',
+      message: 'Error: id is invalid or missing.',
     });
   }
 };
